played with reply and manipulate security

+++ topic/delete.php
@@ -11,10 +11,39 @@
     if(isset($_GET['id'])) {
         $id = $_GET['id'] ; 
 
-        $delete = $conn->query("DELETE FROM topics WHERE id = '$id' ");
-        $delete->execute();
 
-        header("Location: " . APPURL . "/index.php");
+
+
+
+        
+
+        if (isset($_GET['id'])) {
+            $id = $_GET['id'];
+            $select = $conn->prepare("SELECT * FROM topics WHERE id = '$id'");
+            $select->execute();
+                    
+            $topic = $select->fetch(PDO::FETCH_OBJ);
+
+
+			if ($topic->author != $_SESSION['username']) {
+				echo "<script>alert('You are not authorized to delete this topic')</script>";
+                sleep(2);
+				header("Location: " . APPURL . "");
+				exit();
+			}else {
+
+                $delete = $conn->query("DELETE FROM topics WHERE id = '$id' ");
+                $delete->execute();
+        
+                header("Location: " . APPURL . "/index.php");
+
+            }
+
+        }
+
+
+
+       
            
     }
